refactor(cli): extract login and create to actions
